Create CouchbaseContainerConfiguration
