+++ couchbase-driver/src/test/java/org/jnosql/diana/couchbase/document/CouchbaseDocumentCollectionManagerTest.java
@@ -41,8 +41,8 @@ import org.jnosql.diana.api.document.Documents;
 import org.jnosql.diana.api.document.query.DocumentQueryBuilder;
 import org.jnosql.diana.api.key.BucketManager;
 import org.jnosql.diana.api.key.BucketManagerFactory;
-import org.jnosql.diana.couchbase.CouchbaseDocumentTcConfiguration;
-import org.jnosql.diana.couchbase.CouchbaseKeyValueTcConfiguration;
+import org.jnosql.diana.couchbase.configuration.CouchbaseDocumentTcConfiguration;
+import org.jnosql.diana.couchbase.configuration.CouchbaseKeyValueTcConfiguration;
 import org.jnosql.diana.couchbase.CouchbaseUtil;
 import org.jnosql.diana.couchbase.key.CouchbaseKeyValueConfiguration;
 import org.junit.jupiter.api.AfterAll;
