+++ couchbase-driver/src/test/java/org/jnosql/diana/couchbase/document/DocumentQueryTest.java
@@ -20,8 +20,8 @@ import org.jnosql.diana.api.document.DocumentEntity;
 import org.jnosql.diana.api.document.DocumentQuery;
 import org.jnosql.diana.api.key.BucketManager;
 import org.jnosql.diana.api.key.BucketManagerFactory;
-import org.jnosql.diana.couchbase.CouchbaseDocumentTcConfiguration;
-import org.jnosql.diana.couchbase.CouchbaseKeyValueTcConfiguration;
+import org.jnosql.diana.couchbase.configuration.CouchbaseDocumentTcConfiguration;
+import org.jnosql.diana.couchbase.configuration.CouchbaseKeyValueTcConfiguration;
 import org.jnosql.diana.couchbase.CouchbaseUtil;
 import org.jnosql.diana.couchbase.key.CouchbaseKeyValueConfiguration;
 import org.junit.jupiter.api.AfterAll;
