+++ couchbase-driver/src/test/java/org/jnosql/diana/couchbase/key/CouchbaseBucketManagerTest.java
@@ -18,7 +18,7 @@ import org.jnosql.diana.api.Value;
 import org.jnosql.diana.api.key.BucketManager;
 import org.jnosql.diana.api.key.BucketManagerFactory;
 import org.jnosql.diana.api.key.KeyValueEntity;
-import org.jnosql.diana.couchbase.CouchbaseKeyValueTcConfiguration;
+import org.jnosql.diana.couchbase.configuration.CouchbaseKeyValueTcConfiguration;
 import org.jnosql.diana.couchbase.CouchbaseUtil;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
