+++ couchbase-driver/src/test/java/org/jnosql/diana/couchbase/key/CouchbaseMapTest.java
@@ -14,10 +14,9 @@
  */
 package org.jnosql.diana.couchbase.key;
 
-import org.hamcrest.Matchers;
 import org.jnosql.diana.api.key.BucketManager;
 import org.jnosql.diana.api.key.BucketManagerFactory;
-import org.jnosql.diana.couchbase.CouchbaseKeyValueTcConfiguration;
+import org.jnosql.diana.couchbase.configuration.CouchbaseKeyValueTcConfiguration;
 import org.jnosql.diana.couchbase.CouchbaseUtil;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
